Split off the VaccinationIntervention class into a seperate file
