+++ src/intervention_model/VaccinationIntervention.h
@@ -4,27 +4,15 @@
   * SPDX-License-Identifier: MIT
   */
 
-#ifndef INTERVENTION_MODEL_INTERVENTIONS_H_
-#define INTERVENTION_MODEL_INTERVENTIONS_H_
+#ifndef INTERVENTION_MODEL_VACCINATIONINTERVENTION_H_
+#define INTERVENTION_MODEL_VACCINATIONINTERVENTION_H_
 
+#include "Intervention.h"
 #include "AttributeTable.h"
 #include "../protobuf/interventions.pb.h"
 #include "../readers/DataInterface.h"
 
 #include "charm++.h"
-
-class Intervention : public PUP::able {
- public:
-  virtual bool test(const DataInterface &p,
-      std::default_random_engine *generator) const;
-  virtual void apply(DataInterface *p) const;
-  virtual void pup(PUP::er &p);  // NOLINT(runtime/references)
-
-  PUPable_decl(Intervention);
-  Intervention() {}
-  explicit Intervention(CkMigrateMessage *m) :
-    PUP::able(m) {}  // NOLINT(runtime/references)
-};
 
 class VaccinationIntervention : public Intervention {
  public:
@@ -45,4 +33,5 @@ class VaccinationIntervention : public Intervention {
   VaccinationIntervention();
   explicit VaccinationIntervention(CkMigrateMessage *m) : Intervention(m) {}
 };
-#endif  // INTERVENTION_MODEL_INTERVENTIONS_H_
+
+#endif  // INTERVENTION_MODEL_VACCINATIONINTERVENTION_H_
